Update for Lizard API v4
Changed the module the work with API v4. Also changed the tests. Not sure about how to improve the loggin/printing of response urls. @reinout any suggestions? Raise for status is not enough, it doesn't give the details of why the url doesn;t work.  so I added a print_details function.